Вынес код который ищет View по пространствам имен в отдельный класс.
Старая реализация была завязана на Tdi.
Для использования без TDI она не подходит, плюс старая, по смыслу делает другое.
Моя новая предназначена для создания виджета View, на базе Gtk, для ViewModel. Старая знает про Tdi и может вернуть саму вкладку, если она уже Widget. При разделении View и ViewModel такого быть не может.
