+++ QS.Project.Gtk/Views/Resolve/IGtkViewResolver.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using Gtk;
+using QS.ViewModels;
+
+namespace QS.Views.Resolve
+{
+	public interface IGtkViewResolver
+	{
+		Widget Resolve(DialogViewModelBase viewModel);
+	}
+}
